requirements.txt
torch
torchvision
torch_fidelity
tqdm
numpy
Pillow
matplotlib
black
ruff